move augmentHttpRequestOptions to utils
