+++ lib/index.js
@@ -217,17 +217,6 @@ function _transformSnapshot(fields, symbols, data) {
     .value();
 }
 
-function augmentHttpRequestOptions(optionalOptions) {
-  const finalOptions = optionalOptions ? _.clone(optionalOptions) : {};
-
-  if (finalOptions.jar)
-    throw new Error("node-yahoo-finance does not support 'jar' key in " +
-      "optionalHttpRequestOptions, since we need to use our own cookiejar.");
-
-  finalOptions.jar = _cookies.jar;
-  return finalOptions;
-}
-
 function historical(options, optionalHttpRequestOptions, cb) {
   options = _.clone(options);
   _sanitizeHistoricalOptions(options);
@@ -237,8 +226,6 @@ function historical(options, optionalHttpRequestOptions, cb) {
     cb = optionalHttpRequestOptions;
     optionalHttpRequestOptions = undefined;
   }
-
-  var finalHttpRequestOptions = augmentHttpRequestOptions(optionalHttpRequestOptions);
 
   return _cookies.getCrumb(symbols[0])
     .then(function(crumb) {
@@ -250,7 +237,7 @@ function historical(options, optionalHttpRequestOptions, cb) {
           interval: options.period,
           events: options.events,
           crumb: crumb
-        }, finalHttpRequestOptions)
+        }, optionalHttpRequestOptions)
         .then(_utils.parseCSV)
         .then(function (data) {
           return _transformHistorical(symbol, data);
@@ -284,11 +271,10 @@ function snapshot(options, optionalHttpRequestOptions, cb) {
 
   if(optionalHttpRequestOptions && typeof optionalHttpRequestOptions == 'function') {
     cb = optionalHttpRequestOptions;
-    optionalHttpRequestOptions = undefined;
+    optionalHttpRequestOptions = {};
   }
 
-  var finalHttpRequestOptions = augmentHttpRequestOptions(optionalHttpRequestOptions);
-  finalHttpRequestOptions.json = true;
+  optionalHttpRequestOptions.json = true;
 
   if (symbols.length > 1)
     throw new Error("TODO multi symbol support, requires multiple requests in new API");
@@ -304,7 +290,7 @@ function snapshot(options, optionalHttpRequestOptions, cb) {
           // earnings, price, summaryDetail, defaultKeyStatistics, calendarEvents
           modules: 'price,summaryDetail',
           corsDomain: 'finance.yahoo.com'
-        }, finalHttpRequestOptions)
+        }, optionalHttpRequestOptions)
         .then(function(result) {
           var quoteSummary = result.quoteSummary;
           if (!quoteSummary || quoteSummary.error)
