+++ lib/utils.js
@@ -6,6 +6,8 @@ var S = require('string');
 var debug = require('debug')('yahoo-finance:utils');
 var request = require('request-promise');
 var moment = require('moment-timezone');
+var _cookies = require('./financeCookies');
+
 var dateFormats = ['YYYY-MM-DD', 'MM/DD/YYYY'];
 
 function camelize(text) {
@@ -15,9 +17,22 @@ function camelize(text) {
     .s;
 }
 
+function augmentHttpRequestOptions(optionalOptions) {
+  const finalOptions = optionalOptions ? _.clone(optionalOptions) : {};
+
+  if (finalOptions.jar)
+    throw new Error("node-yahoo-finance does not support 'jar' key in " +
+      "optionalHttpRequestOptions, since we need to use our own cookiejar.");
+
+  finalOptions.jar = _cookies.jar;
+  return finalOptions;
+}
+
 function download(uri, qs, optionalHttpRequestOptions) {
+  const finalHttpRequestOptions = augmentHttpRequestOptions(optionalHttpRequestOptions);
+
   debug(url.format({pathname: uri, query: qs}));
-  return request(_.extend({uri: uri, qs: qs}, optionalHttpRequestOptions));
+  return request(_.extend({uri: uri, qs: qs}, finalHttpRequestOptions));
 }
 
 function parseCSV(text) {
